# [symbolic] Speed up matmul