Towards https://github.com/RobotLocomotion/drake/issues/19202 and https://github.com/RobotLocomotion/drake/issues/10900.

The crux of the improvement here is `GenericGevv` computing the inner product directly using `AddFactory`, without creating intermediate Expression nodes.

The calculation of `MatrixX<double> * MatrixX<Variable>` can still be _much_ better optimized.  That will come in a future commit (#19345).

---

For multiplying two matrices each with shape=(50, 50), this yields ~4x throughput compared to master (data shown below).  For shape=(200, 200), the throughput yield is more like ~11x.

On master:
```
------------------------------------------------------------
Benchmark                  Time             CPU   Iterations
------------------------------------------------------------
GemmDV/50_mean           225 ms          225 ms           10
GemmDE/50_mean          82.0 ms         82.0 ms           10
GemmEE/50_mean           151 ms          151 ms           10
```

On this branch:
```
------------------------------------------------------------
Benchmark                  Time             CPU   Iterations
------------------------------------------------------------
GemmDV/50_mean          63.9 ms         63.9 ms           10
GemmDE/50_mean          22.6 ms         22.6 ms           10
GemmEE/50_mean          43.4 ms         43.4 ms           10
```

The benchmark program is found at #19303.

---

I've successfully tested this against Anzu CI.

<!-- Reviewable:start -->
- - -
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/RobotLocomotion/drake/19304)
<!-- Reviewable:end -->